Rename plateFunctions to toolComponentFunctions
