+++ ui/src/KaptureApp/api/kapture.js
@@ -7,7 +7,7 @@ import {
   COMPONENT_TYPE_MEDIUM,
   COMPONENT_TYPE_SUPPLEMENT,
 } from 'KaptureApp/config/componentTypes';
-import { createComponent } from 'KaptureApp/utils/plateFunctions';
+import { createComponent } from 'KaptureApp/utils/toolComponentFunctions';
 
 export function searchActivities(page, size, nameContains, descContains) {
   let queryString = '';
