+++ ui/src/KaptureApp/utils/toolComponentFunctions.js
